+++ kikuchipy/data/__init__.py
@@ -22,8 +22,8 @@ Some datasets must be downloaded from the web. For more test datasets,
 see :doc:`open datasets <open_datasets>`.
 """
 
-from pathlib import Path
 import os
+from pathlib import Path
 from typing import Union
 
 import pooch as ppooch
@@ -111,7 +111,7 @@ def nickel_ebsd_small(**kwargs) -> EBSD:
     signal : EBSD
         EBSD signal.
     """
-    return _load(filename="data/kikuchipy/patterns.h5", **kwargs)
+    return _load(filename="data/kikuchipy_h5ebsd/patterns.h5", **kwargs)
 
 
 def nickel_ebsd_master_pattern_small(**kwargs) -> EBSDMasterPattern: