+++ kikuchipy/io/plugins/nordif_calibration_patterns.py
@@ -118,7 +118,7 @@ def _get_coordinates(filename: str) -> List[Tuple[int, int]]:
         raise ValueError(err)
     xy = []
     for line in content[l_start + 1 :]:
-        match = re.search("Calibration \((.*)\)", line)
+        match = re.search(r"Calibration \((.*)\)", line)
         try:
             match = match.group(1)
             match = match.split(",")